refactor(day06-oop): 重命名继承方法重写示例代码的包名
