+++ day06-oop/src/com/itheima/_05继承方法重写/Demo05.java
@@ -1,4 +1,4 @@
-package com.itheima._5继承方法重写;
+package com.itheima._05继承方法重写;
 
 /**
  * @Description Demo05
